Added application of AC to the Virtual Engineering of biofuels.
Fixed plot_nd to compute distance in parameter space where each dimension is normalized by the range of its xlimits.
Changed the minimizer to be 'Powell' instead of 'SLSQP' which is better if the optimum lies on a parameter's allowable bounds.
Updated the conda environment file to run AC and VE without warnings.