Added option to write csv of all data computed. Fixed  bug for multifidelity with mixed types.
One output file is written per fidelity level.
The filenames are specified with options.output_data_filenames.
Fixed a bug for multifidelity with mixed types: was always using the HF simulation.
Added feature: If the objective function is not evaluated in the input file
(i.e., if only parameters are given), then the objective function is computed.