For multifidelity, at every point in the design space where a simulation is performed, compute all lower fidelity level simulations there too.
This is controlled by options.perform_lower_sims which is True by default.
Added a warning if HF has more n_init than LF. (This is typically leads to poor perforamce of BayesOpt and generally a poor allocation of resources)
Support multifidelity with n_init_samp = 0 for level zero but nonzero for higher levels. Previously this was not permitted.