Minimize continuous parameters using scipy.minimize and discrete params using scipy.brute.
Most methods in scipy.minimize don't support interger parameters.
So we swich to use scipy.brute, which is an exhaustive search over
discrete parameters. This can be efficient for discrete
types as long as there is a modest number of permitted values.
The Model object constructor requires the parameters in the params
list to be ordered so that continuous variables appear before
any categorical or ordered types. See the README for details.