Support simulations with multiple outputs.
Although the simulations has multiple outputs, and these are stored in the dataset object, the surrogate model
is still single output, so the user specifies which output the model is trained on.
Renamed the variable n_dim -> n_in to be more specific
Added the variable n_out for the number of outputs.
User should specify n_out when creating DataSet. Default to 1.
User should specify i_out when creating surrogate. Default to 0.