Modularize add_bo_samples to call get_bo_sample in a loop.
get_bo_sample gets the x coordinates that minimize the acquisition function. It only returns 1 sample.
It does not allocate memory in the dataset (no masking), does not run simulations, and does not interact with HERO.
add_bo_samples does do all of these things and retrains the surrogate. It has an outer loop that does all these
operations for each sample in the batch size.