Changing to static variance threshold and use Python 3.11.
Removed dynamic variance threshold to simplify debugging.
README updated to require Python <3.12 since SMT 1.3 imports imp,
which is not supported with Python 3.12. With Python 3.12, on Kestrel
this code was not able to call SMT.
Fixed a bug where the surrogate was not being retrained.
add_xnum_sample is called with the surrogate argument so that retraining occurs.
Fixed memory bug where PyFloat was being freed before it is used.